requirements.txt
asn1crypto==0.24.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
blinker==1.4
bs4==0.0.1
cffi==1.11.5
cheroot==6.5.4
Click==7.0
cryptography==2.4.2
cycler==0.10.0
ecdsa==0.13
filelock==3.0.10
Flask==1.0.2
Flask-Mail==0.9.1
future==0.17.1
gunicorn==19.9.0
httplib2==0.12.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
lxml==4.3.0
Markdown==2.6.11
MarkupSafe==1.1.0
matplotlib==3.0.2
more-itertools==5.0.0
mpmath==1.1.0
numpy==1.15.4
oauth2==1.9.0.post1
oauthlib==2.1.0
ply==3.11
pyasn1==0.4.5
pycparser==2.19
PyLTI==0.7.0
pyparsing==2.3.0
python-dateutil==2.7.5
python-dotenv==0.10.1
python-jose==3.0.1
rsa==4.0
scipy==1.2.0
six==1.12.0
soupsieve==1.6.2
SQLAlchemy==1.2.15
uuid==1.30
uWSGI==2.0.17.1
virtualenv==16.1.0
websockets==7.0
Werkzeug==0.14.1
